Basic player movement script added
Added a test scene and basic player movement script. Tested and worked on my android and PC.  Only jumping works on Android for now because the lack of touch control buttons but left right movement is available on PC with the arrow keys. I don't know if this is the jumping control scheme we want; when I was thinking about it I figured this would method would be good for touch controls but if we decide to change it, it should be easy to change. If we do go with this drag and release style it would have to be changed to only start when a certain area of the screen is touched - I'm guessing around the player - as we will need the rest of the screen space for other ui/control elements.