Finish Trajectory + Fixes
Trajectory arc is now complete and hiding it has been fixed. My last commit had hiding working but the arc would jerk around whenever appearing. The method in this commit is much better and simpler and gets rid of the jerkyness. Also fixed some minor things with movement/jumping and the bug where the player would hang midair if the screen was touched while in the air.