Added keycard objects
When the keycard is picked up it's name is put into the hashset in the simpleplayermovement script (this can be changed to have the hashset in the player script or whatever later) and is destroyed. You would then probably need to check  in the door code if the items hashset contains the required keycards using items.Contains("blueKeycard") or something.